requirements.txt
anyio==4.4.0
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
ecdsa==0.19.0
fastapi==0.95.2
h11==0.14.0
idna==3.7
motor==3.2.0
passlib==1.7.4
pyasn1==0.6.0
pydantic==1.10.17
pymongo==4.8.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.22.0
fastapi==0.95.2
uvicorn==0.22.0
pymongo
motor==3.2.0
bcrypt==4.0.1
python-jose==3.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
